feat(android-needs-review): Rename AutomatedChecks* to Results* in e2e tests
The idea is to separate test code for the results view -- which covers the overall content of the app such as the command bar, left nav bar, and page content -- into its own file, while keeping the test code for the automated checks view -- the automated test results -- in its own file.

- Rename `ElectronAutomatedChecksView` to `ElectronResultsView` in `TelemetryEventSource`
- Extract code having to do with results view from `AutomatedChecksViewController` into `ResultsViewController` (likewise with selectors)
- Extract tests having to do with results view into a separate e2e test file
- Both sets of tests should still test for accessibility
