+++ src/tests/electron/common/view-controllers/automated-checks-view-controller.ts
@@ -1,12 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { LeftNavItemKey } from 'electron/types/left-nav-item-key';
+import { ScreenshotViewSelectors } from 'tests/electron/common/element-identifiers/screenshot-view-selectors';
 import { SpectronAsyncClient } from 'tests/electron/common/view-controllers/spectron-async-client';
 import { settingsPanelSelectors } from 'tests/end-to-end/common/element-identifiers/details-view-selectors';
-import {
-    AutomatedChecksViewSelectors,
-    ScreenshotViewSelectors,
-} from '../element-identifiers/automated-checks-view-selectors';
+import { AutomatedChecksViewSelectors } from '../element-identifiers/automated-checks-view-selectors';
 import { ViewController } from './view-controller';
 
 export class AutomatedChecksViewController extends ViewController {
@@ -32,21 +29,6 @@ export class AutomatedChecksViewController extends ViewController {
         await this.client.click(selector);
     }
 
-    public async waitForViewVisible(): Promise<void> {
-        await this.waitForSelector(AutomatedChecksViewSelectors.mainContainer);
-    }
-
-    public async waitForScreenshotViewVisible(): Promise<void> {
-        await this.waitForSelector(ScreenshotViewSelectors.screenshotView);
-    }
-
-    public async openSettingsPanel(): Promise<void> {
-        await this.waitForSelector(AutomatedChecksViewSelectors.settingsButton);
-        await this.click(AutomatedChecksViewSelectors.settingsButton);
-        await this.waitForSelector(settingsPanelSelectors.settingsPanel);
-        await this.waitForMilliseconds(750); // Allow for fabric's panel animation to settle
-    }
-
     public async setToggleState(toggleSelector: string, newState: boolean): Promise<void> {
         await this.waitForSelector(toggleSelector);
         const oldState = await this.client.getAttribute(toggleSelector, 'aria-checked');
@@ -64,15 +46,5 @@ export class AutomatedChecksViewController extends ViewController {
             : settingsPanelSelectors.disabledToggle(toggleSelector);
 
         await this.waitForSelector(toggleInStateSelector);
-    }
-
-    public async clickLeftNavItem(key: LeftNavItemKey): Promise<void> {
-        const selector = this.getSelectorForLeftNavItemLink(key);
-        await this.waitForSelector(selector);
-        await this.click(selector);
-    }
-
-    private getSelectorForLeftNavItemLink(key: LeftNavItemKey): string {
-        return `${AutomatedChecksViewSelectors.leftNav} a[data-automation-id="${key}"]`;
     }
 }
